requirements.txt
-e .
aws-cdk.aws-iam
aws-cdk.aws-codepipeline
aws-cdk.aws-codebuild
 
